Time: 4 ms (71.12%), Space: 15.8 MB (27.27%) - LeetHub

+++ 1643-kth-smallest-instructions/1643-kth-smallest-instructions.cpp
@@ -14,7 +14,7 @@ public:
         {
             if(currH>0)
             {
-                double ifH = fac[total-1]/fac[currH-1]/fac[currV];
+                long long ifH = fac[total-1]/fac[currH-1]/fac[currV];
                 if(ifH>=k)
                 {
                     a+="H";
